requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-debug
flask-debugtoolbar
flask-wtf
peewee
flask-security
bcrypt
